requirements.txt
tensorflow[and-cuda]
pillow
scikit-learn
scikit-image
segmentation-models
matplotlib